// nextjs/src/api/payload-client/model/communities.ts
/**
 * Generated by orval v7.1.1 🍺
 * Do not edit manually.
 * payload
 * A blank template to get started with Payload
 * OpenAPI spec version: 0.0.1
 */
import type { _CommunitiesStatus } from './_communitiesStatus';
import type { CommunitiesCommunityMembersItem } from './communitiesCommunityMembersItem';
import type { CommunitiesCommunityOrganizersItem } from './communitiesCommunityOrganizersItem';
import type { CommunitiesCreatedBy } from './communitiesCreatedBy';
import type { CommunitiesParentCommunity } from './communitiesParentCommunity';
import type { CommunitiesSubcommunitiesItem } from './communitiesSubcommunitiesItem';

export interface Communities {
  /** @nullable */
  _status?: _CommunitiesStatus;
  /** @nullable */
  communityLongDescription?: string | null;
  /** @nullable */
  communityMembers?: CommunitiesCommunityMembersItem[] | null;
  communityName: string;
  /** @nullable */
  communityOrganizers?: CommunitiesCommunityOrganizersItem[] | null;
  communitySummary: string;
  createdAt: string;
  createdBy?: CommunitiesCreatedBy;
  parentCommunity?: CommunitiesParentCommunity;
  slug: string;
  /** @nullable */
  subcommunities?: CommunitiesSubcommunitiesItem[] | null;
  updatedAt: string;
}